app/src/androidTest/java/io/github/kiyohitonara/souji/data/AppInfoDeviceDataSourceTest.kt
/*
 * Copyright (c) 2024 Kiyohito Nara
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

package io.github.kiyohitonara.souji.data

import android.content.Context
import android.content.pm.PackageManager
import androidx.test.core.app.ApplicationProvider
import androidx.test.ext.junit.runners.AndroidJUnit4
import kotlinx.coroutines.flow.last
import kotlinx.coroutines.runBlocking
import org.junit.Assert.assertTrue
import org.junit.Before
import org.junit.Test
import org.junit.runner.RunWith

@RunWith(AndroidJUnit4::class)
class AppInfoDeviceDataSourceTest {
    private lateinit var dataSource: AppInfoDeviceDataSource
    private lateinit var packageManager: PackageManager

    @Before
    fun setup() {
        val context = ApplicationProvider.getApplicationContext<Context>()
        packageManager = context.packageManager
        dataSource = AppInfoDeviceDataSource(context)
    }

    @Test
    fun testGetApps() = runBlocking {
        val apps = dataSource.getApps().last()

        assertTrue("apps size is ${apps.size}.", apps.isNotEmpty())
    }

    @Test(expected = UnsupportedOperationException::class)
    fun testUpsertApp() = runBlocking {
        val appInfo = dataSource.getApps().last().first()

        dataSource.upsertApp(appInfo)
    }
}